useDepartmentStore.test.ts
import { mockNuxtImport } from '@nuxt/test-utils/runtime';
import { afterAll, beforeAll, describe, expect, it, vi } from 'vitest';
import type { PaginationFilter } from '~/interfaces/common/pagination-filter';

const pagination = ref<PaginationFilter>(defaultPaginationValues);
const mockDepartmentData = [
  { id: 1, departmentCode: 'Dep1', departmentName: 'dep of hike' },
];
const mockedSetPaginationDetails = vi.fn();

describe('useDepartmentStore', () => {
  beforeAll(() => {
    mockNuxtImport(
      'usePagination',
      vi.fn().mockReturnValue(() => ({
        pagination: pagination,
        setPaginationDetails: mockedSetPaginationDetails,
      }))
    );

    mockNuxtImport(
      'useFetchData',
      vi.fn().mockReturnValue(() => ({
        fetchData: vi.fn().mockResolvedValue({
          data: mockDepartmentData,
          pagination: {
            page: 1,
            perPage: 5,
            totalCount: 0,
            searchString: '',
          },
        }),
      }))
    );

    mockNuxtImport(
      'useCreateData',
      vi.fn().mockReturnValue(() => ({
        createData: vi.fn().mockResolvedValue({ id: 3 }),
      }))
    );

    mockNuxtImport(
      'useUpdateData',
      vi.fn().mockReturnValue(() => ({
        updateData: vi.fn().mockResolvedValue({ id: 3 }),
      }))
    );

    mockNuxtImport(
      'useDeleteData',
      vi.fn().mockReturnValue(() => ({
        deleteData: vi.fn().mockResolvedValue({ id: 3 }),
      }))
    );

    vi.mock('primevue/usetoast', () => ({
      useToast: () => ({
        add: vi.fn(),
      }),
    }));
  });

  afterAll(() => {
    vi.clearAllMocks();
    vi.resetAllMocks();
  });

  it('Should fetch all departments', async () => {
    const store = useDepartmentStore();
    await store.fetchDepartments();

    expect(store.departmentList).toEqual(mockDepartmentData);
    expect(mockedSetPaginationDetails).toBeCalledWith('', 1, 5, 0);
  });

  it('Should add department details', async () => {
    const store = useDepartmentStore();
    await store.createNewDepartment({
      departmentCode: 'Dep2',
      departmentName: 'dep of swim',
    });

    expect(store.departmentList).toEqual(mockDepartmentData);
    expect(store.isOperationSuccessful).toEqual(true);
    expect(mockedSetPaginationDetails).toBeCalled();
  });

  it('Should update department details', async () => {
    const store = useDepartmentStore();
    await store.updateDepartment({
      id: 1,
      departmentCode: 'Dep1',
      departmentName: 'dep of hiking',
    });

    expect(store.departmentList).toEqual(mockDepartmentData);
    expect(store.isOperationSuccessful).toEqual(true);
    expect(mockedSetPaginationDetails).toBeCalled();
  });

  it('Should delete department details', async () => {
    const store = useDepartmentStore();
    await store.deleteDepartment(3);

    expect(store.departmentList).toEqual(mockDepartmentData);
    expect(store.isOperationSuccessful).toEqual(true);
    expect(mockedSetPaginationDetails).toBeCalled();
  });
});
